fix: ensure es and kibana data directories exist
This ensures that the directories for Docker data for kibana and es
exist on the host.

The kibana and es process will show an error otherwise and refuse to
boot. Example from the kibana process:

FATAL  Error: Unable to write to UUID file at
/usr/share/kibana/data/uuid. Ensure Kibana has sufficient permissions to
read / write to this file.  Error was: EACCES